Remove usage of collection classes
