+++ src/TaskConfiguration/__init__.py
@@ -1,15 +1,11 @@
-from .task_category_collection import TaskCategoryCollection
 from .task_category import TaskCategory
 from .task_category_factory import TaskCategoryFactory
-from .task_category_collection_factory import TaskCategoryCollectionFactory
 from .task_configuration import TaskConfiguration
 from .task_configuration_factory import TaskConfigurationFactory
 
 __all__ = [
-    "TaskCategoryCollection",
     "TaskCategory",
     "TaskCategoryFactory",
-    "TaskCategoryCollectionFactory",
     "TaskConfiguration",
     "TaskConfigurationFactory"
 ]
