--- src/TaskConfiguration/task_category_collection.py
@@ -1,4 +1,0 @@
-class TaskCategoryCollection:
-
-    def __init__(self, categories):
-        self.categories = categories
