--- src/TaskConfiguration/task_configuration_collection.py
@@ -1,4 +1,0 @@
-class TaskConfigurationCollection:
-
-    def __init__(self, configurations):
-        self.configurations = configurations
