+++ src/main.py
@@ -3,7 +3,8 @@ from pathlib import Path
 
 from src.Embeddings.embeddings_factory import EmbeddingsFactory
 from src.EntityLinking.entity_linkings_factory import EntityLinkingsFactory
-from src.TaskConfiguration import TaskCategoryCollectionFactory
+from src.TaskConfiguration import TaskConfigurationFactory, TaskCategoryFactory
+from src.TestConfiguration.test_configuration import TestConfiguration
 
 
 def main(args):
