+++ test/TaskConfiguration/test_task_configuration_factory.py
@@ -5,6 +5,31 @@ from test.base_test_case import BaseTestCase
 
 class TestTaskConfigurationFactory(BaseTestCase):
 
+    def test_create_configurations_from_file(self):
+        file = self._random_test_file()
+        with open(file, "w+") as file_output:
+            print("""
+            configuration:
+                task_configurations:
+                    - task_configuration:
+                        type: analogy
+                        enabled: true
+                    - task_configuration:
+                        type: neighborhood
+                        enabled: false""",
+                  file=file_output)
+
+        task_configurations = TaskConfigurationFactory.create_configurations_from_file(file)
+        self.assertEqual(2, len(task_configurations))
+
+        config = task_configurations[0]
+        self.assertEqual(TaskType.ANALOGY, config.task_type)
+        self.assertTrue(config.enabled)
+
+        config = task_configurations[1]
+        self.assertEqual(TaskType.NEIGHBORHOOD, config.task_type)
+        self.assertFalse(config.enabled)
+
     def test_extract_enabled(self):
         self.assertTrue(TaskConfigurationFactory._extract_enabled({"enabled": "true"}))
         self.assertTrue(TaskConfigurationFactory._extract_enabled({"enabled": "TRUE"}))
